refactor: extract shared logic to GameBoard
